requirement.txt
django
djongo
djangorestframework
djoser
pika